LayoutTests/fast/dom/Window/window-property-shadowing_-expected.txt

PASS: locationbar is not shadowed
PASS: menubar is not shadowed
PASS: personalbar is not shadowed
PASS: statusbar is not shadowed
PASS: toolbar is not shadowed
PASS: devicePixelRatio is not shadowed
PASS: defaultStatus is not shadowed
PASS: defaultstatus is not shadowed
PASS: status is not shadowed
PASS: frames is not shadowed
PASS: event is not shadowed
PASS: innerHeight is not shadowed
PASS: innerWidth is not shadowed
PASS: length is not shadowed
PASS: name is not shadowed
PASS: navigator is not shadowed
PASS: clientInformation is not shadowed
PASS: offscreenBuffering is not shadowed
PASS: opener is not shadowed
PASS: outerHeight is not shadowed
PASS: outerWidth is not shadowed
PASS: parent is not shadowed
PASS: screenX is not shadowed
PASS: screenY is not shadowed
PASS: screenLeft is not shadowed
PASS: screenTop is not shadowed
PASS: scrollX is not shadowed
PASS: scrollY is not shadowed
PASS: self is not shadowed
PASS: top is not shadowed
PASS: getSelection is not shadowed
PASS: getComputedStyle is not shadowed
PASS: getMatchedCSSRules is not shadowed
PASS: atob is not shadowed
PASS: btoa is not shadowed
PASS: scroll is not shadowed
PASS: scrollBy is not shadowed
PASS: scrollTo is not shadowed
PASS: moveBy is not shadowed
PASS: moveTo is not shadowed
PASS: resizeBy is not shadowed
PASS: resizeTo is not shadowed
PASS: alert is not shadowed
PASS: confirm is not shadowed
PASS: prompt is not shadowed
PASS: open is not shadowed
PASS: print is not shadowed
PASS: setTimeout is not shadowed
PASS: clearTimeout is not shadowed
PASS: focus is not shadowed
PASS: blur is not shadowed
PASS: close is not shadowed
PASS: setInterval is not shadowed
PASS: clearInterval is not shadowed
PASS: captureEvents is not shadowed
PASS: releaseEvents is not shadowed
PASS: addEventListener is not shadowed
PASS: removeEventListener is not shadowed
PASS: accessing showModalDialog resulted in an exception. That's because the client (probably DumpRenderTree) cannot run modal - and showModalDialog isn't shadowed.
PASS: find is not shadowed
PASS: stop is not shadowed
PASS: Image is not shadowed
PASS: Option is not shadowed
PASS: DOMException is not shadowed
PASS: XMLHttpRequest is not shadowed
PASS: XSLTProcessor is not shadowed
PASS: StyleSheet is not shadowed
PASS: CSSStyleSheet is not shadowed
PASS: CSSValue is not shadowed
PASS: CSSPrimitiveValue is not shadowed
PASS: CSSValueList is not shadowed
PASS: CSSRule is not shadowed
PASS: CSSCharsetRule is not shadowed
PASS: CSSFontFaceRule is not shadowed
PASS: CSSImportRule is not shadowed
PASS: CSSMediaRule is not shadowed
PASS: CSSPageRule is not shadowed
PASS: CSSStyleRule is not shadowed
PASS: CSSStyleDeclaration is not shadowed
PASS: MediaList is not shadowed
PASS: Counter is not shadowed
PASS: CSSRuleList is not shadowed
PASS: Rect is not shadowed
PASS: DOMImplementation is not shadowed
PASS: DocumentFragment is not shadowed
PASS: Document is not shadowed
PASS: Node is not shadowed
PASS: NodeList is not shadowed
PASS: NamedNodeMap is not shadowed
PASS: CharacterData is not shadowed
PASS: Attr is not shadowed
PASS: Element is not shadowed
PASS: Text is not shadowed
PASS: Comment is not shadowed
PASS: CDATASection is not shadowed
PASS: DocumentType is not shadowed
PASS: Notation is not shadowed
PASS: Entity is not shadowed
PASS: EntityReference is not shadowed
PASS: ProcessingInstruction is not shadowed
PASS: HTMLDocument is not shadowed
PASS: HTMLElement is not shadowed
PASS: HTMLAnchorElement is not shadowed
PASS: HTMLAppletElement is not shadowed
PASS: HTMLAreaElement is not shadowed
PASS: HTMLBRElement is not shadowed
PASS: HTMLBaseElement is not shadowed
PASS: HTMLBaseFontElement is not shadowed
PASS: HTMLBlockquoteElement is not shadowed
PASS: HTMLBodyElement is not shadowed
PASS: HTMLButtonElement is not shadowed
PASS: HTMLCanvasElement is not shadowed
PASS: HTMLDListElement is not shadowed
PASS: HTMLDirectoryElement is not shadowed
PASS: HTMLDivElement is not shadowed
PASS: HTMLEmbedElement is not shadowed
PASS: HTMLFieldSetElement is not shadowed
PASS: HTMLFontElement is not shadowed
PASS: HTMLFormElement is not shadowed
PASS: HTMLFrameElement is not shadowed
PASS: HTMLFrameSetElement is not shadowed
PASS: HTMLHRElement is not shadowed
PASS: HTMLHeadElement is not shadowed
PASS: HTMLHeadingElement is not shadowed
PASS: HTMLHtmlElement is not shadowed
PASS: HTMLIFrameElement is not shadowed
PASS: HTMLImageElement is not shadowed
PASS: HTMLInputElement is not shadowed
PASS: HTMLIsIndexElement is not shadowed
PASS: HTMLLIElement is not shadowed
PASS: HTMLLabelElement is not shadowed
PASS: HTMLLegendElement is not shadowed
PASS: HTMLLinkElement is not shadowed
PASS: HTMLMapElement is not shadowed
PASS: HTMLMarqueeElement is not shadowed
PASS: HTMLMenuElement is not shadowed
PASS: HTMLMetaElement is not shadowed
PASS: HTMLModElement is not shadowed
PASS: HTMLOListElement is not shadowed
PASS: HTMLObjectElement is not shadowed
PASS: HTMLOptGroupElement is not shadowed
PASS: HTMLOptionElement is not shadowed
PASS: HTMLParagraphElement is not shadowed
PASS: HTMLParamElement is not shadowed
PASS: HTMLPreElement is not shadowed
PASS: HTMLQuoteElement is not shadowed
PASS: HTMLScriptElement is not shadowed
PASS: HTMLSelectElement is not shadowed
PASS: HTMLStyleElement is not shadowed
PASS: HTMLTableCaptionElement is not shadowed
PASS: HTMLTableCellElement is not shadowed
PASS: HTMLTableColElement is not shadowed
PASS: HTMLTableElement is not shadowed
PASS: HTMLTableRowElement is not shadowed
PASS: HTMLTableSectionElement is not shadowed
PASS: HTMLTextAreaElement is not shadowed
PASS: HTMLTitleElement is not shadowed
PASS: HTMLUListElement is not shadowed
PASS: Event is not shadowed
PASS: MutationEvent is not shadowed
PASS: NodeFilter is not shadowed
PASS: Range is not shadowed
PASS: RangeException is not shadowed
PASS: XMLDocument is not shadowed
PASS: DOMParser is not shadowed
PASS: XMLSerializer is not shadowed
